Form validation and error handling

+++ backend/src/controllers/UrlKeyReuseController.ts
@@ -4,32 +4,44 @@ import { AppDataSource } from "../database";
 import { Url } from "../entities/Url";
 import { EncurtURLRepository } from "../repositories/EncurtURLRepository";
 import { KeyReuseRepository } from "../repositories/KeyReuseRepository";
+import { UrlRepository } from "../repositories/UrlRepository";
 
 export class UrlKeyReuseController {
   handle: RequestHandler = async (req: Request, res: Response) => {
-    const repository = new EncurtURLRepository(
-      Url,
-      new MongoEntityManager(AppDataSource)
-    );
+    let { encurt_url } = req.body;
 
-    try {
-      let { encurt_url } = req.body;
+    const repo = UrlRepository();
 
-      const e_url = await repository.findByEncurtURL(encurt_url);
+    const search = await repo.find({
+      select: { encurt_url: true },
+      where: { encurt_url: `${encurt_url}` },
+    });
 
-      const reuse = e_url.filter(async function recycle(data: Url) {
-        let chave = data.chave;
-        let key_available = true;
+    if (search.find(async (dataEncurt: Url) => dataEncurt.encurt_url)) {
+      const repository = new EncurtURLRepository(
+        Url,
+        new MongoEntityManager(AppDataSource)
+      );
 
-        let keyReuse = KeyReuseRepository().create({ chave, key_available });
-        await KeyReuseRepository().save(keyReuse);
+      try {
+        const e_url = await repository.findByEncurtURL(encurt_url);
 
-        await repository.delete(data.id);
+        const reuse = e_url.filter(async function recycle(data: Url) {
+          let chave = data.chave;
+          let key_available = true;
 
-        return res.json("A URL foi excluída com sucesso!");
-      });
-    } catch (error) {
-      console.error(error);
+          let keyReuse = KeyReuseRepository().create({ chave, key_available });
+          await KeyReuseRepository().save(keyReuse);
+
+          await repository.delete(data.id);
+        });
+
+        return res.json("A URL curta foi excluída com sucesso!");
+      } catch (error) {
+        console.error(error);
+      }
+    } else {
+      return res.json("Ops! Essa URL curta não existe!");
     }
   };
 }
